sis_project/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
openpyxl==3.1.2
gunicorn==21.2.0
django-redis==5.4.0
redis==5.0.1
Pillow==10.0.1
python-docx==1.1.0